feat: Update README with personal information and location
This commit updates the README.md file to include personal information and location. The greeting message has been changed to "😊 Hey! Nice to see you." and additional details about the author have been added. The location is now dynamically generated based on the value provided in the main.go file.